cloudfunctions/register/requirements.txt
google-cloud-datastore>=1.7.1
google-cloud-kms>=1.0.0
marshmallow>=2.18.1
PyJWT>=1.7.1
